requirements.txt
absl-py==2.1.0
accelerate==1.7.0
adaseq==0.6.6
addict==2.4.0
aiofiles==23.2.1
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiosignal==1.3.2
albucore==0.0.16
albumentations==1.4.10
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.8.0
appdirs==1.4.4
archspec==0.2.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
astor==0.8.1
asttokens==3.0.0
astunparse==1.6.3
async-lru==2.0.4
async-timeout==5.0.1
attrs==25.1.0
av==14.4.0
babel==2.17.0
Baker==1.3
basicsr==1.4.2
beautifulsoup4==4.13.3
bezier==2024.6.20
bidict==0.23.1
binaryornot==0.4.4
bitsandbytes==0.42.0
black==25.1.0
bleach==6.2.0
blinker==1.9.0
boltons==23.1.1
Brotli==1.1.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.4.2
click==8.1.8
clip==0.2.0
cloudpickle==3.1.1
colorama==0.4.6
coloredlogs==15.0.1
ConfigArgParse==1.7.1
contourpy==1.3.1
controlnet_aux==0.0.10
cookiecutter==2.6.0
crcmod==1.7
cryptography==43.0.3
cuda-python==12.2.0
cycler==0.12.1
Cython==3.0.12
dataclasses-json==0.6.7
datasets==2.18.0
debugpy==1.8.13
decorator==5.1.1
deepspeed==0.12.6
defusedxml==0.7.1
Deprecated==1.2.18
diffusers==0.34.0
dill==0.3.8
distro==1.8.0
docker-pycreds==0.4.0
easydict==1.13
efficientnet-pytorch==0.7.1
einops==0.7.0
exceptiongroup==1.2.2
executing==2.2.0
fairscale==0.4.13
fastjsonschema==2.21.1
ffmpy==0.5.0
filelock==3.18.0
flash-attn==2.5.2
Flask==3.0.3
Flask-Cors==5.0.0
Flask-Login==0.6.3
flatbuffers==25.2.10
fonttools==4.56.0
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2025.5.1
ftfy==6.3.1
future==1.0.0
fvcore==0.1.5.post20221221
gast==0.6.0
gdown==5.2.0
gevent==25.5.1
geventhttpclient==2.3.4
google-auth==2.38.0
google-auth-oauthlib==1.2.1
google-pasta==0.2.0
gradio==3.44.3
gradio_client==0.5.0
graphql-core==3.2.6
greenlet==3.2.3
groovy==0.1.2
grpcio==1.70.0
gunicorn==23.0.0
h11==0.14.0
h5py==3.13.0
hf-xet==1.1.3
hjson==3.1.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.33.0
humanfriendly==10.0
hydra-core==1.3.2
idna==3.10
ijson==3.4.0
imageio==2.33.1
imagesize==1.4.1
importlib_metadata==8.7.0
importlib_resources==6.5.2
insightface==0.7.3
iopath==0.1.9
ipykernel==6.29.5
ipython==8.33.0
isoduration==20.11.0
itsdangerous==2.2.0
jax==0.6.1
jaxlib==0.6.1
jedi==0.19.2
Jinja2==3.1.6
jmespath==0.10.0
joblib==1.4.2
json5==0.10.0
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
kornia==0.7.1
lazy_loader==0.4
libclang==18.1.1
libmambapy==1.5.5
lmdb==1.6.2
locust==2.37.10
locust-cloud==1.23.2
logzero==1.7.0
mamba==1.5.5
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.26.1
matplotlib==3.7.4
matplotlib-inline==0.1.7
matrix-client==0.4.0
mdurl==0.1.2
mediapipe==0.10.21
menuinst==2.0.1
mistune==3.1.2
mixpanel==4.10.1
ml_dtypes==0.5.1
mpmath==1.3.0
msgpack==1.1.1
multidict==6.1.0
multiprocess==0.70.16
MultiScaleDeformableAttention==1.0
munch==4.0.0
mypy_extensions==1.1.0
narwhals==1.29.1
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
ncnn==1.0.20240820
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.1
notebook_shim==0.2.4
numpy==1.26.4
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-ml-py==12.570.86
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
oauthlib==3.2.2
omegaconf==2.3.0
opencv-contrib-python==4.10.0.84
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
opt-einsum==3.3.0
orjson==3.10.15
oss2==2.19.1
overrides==7.7.0
packaging==25.0
paddlepaddle-gpu==3.0.0b1
pandas==2.1.4
pandocfilters==1.5.1
parso==0.8.4
pathspec==0.12.1
peft==0.15.2
pexpect==4.9.0
pillow==10.4.0
pip==23.3.2
platformdirs==4.3.8
pluggy==1.3.0
portalocker==3.1.1
prettytable==3.15.1
prodigyopt==1.0
prometheus_client==0.21.1
prompt_toolkit==3.0.50
propcache==0.3.0
protobuf==4.25.8
psutil==5.9.8
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==19.0.1
pyarrow-hotfix==0.6
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycocotools==2.0.8
pycosat==0.6.6
pycparser==2.21
pycryptodome==3.21.0
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
PyGithub==2.6.1
Pygments==2.19.1
PyJWT==2.10.1
PyMuPDF==1.25.1
PyNaCl==1.5.0
pynvml==12.0.0
pyparsing==3.2.1
pypdfium2==4.30.1
PySocks==1.7.1
pyspng==0.1.3
python-dateutil==2.9.0.post0
python-engineio==4.12.2
python-json-logger==3.3.0
python-multipart==0.0.20
python-slugify==8.0.4
python-socketio==5.13.0
pytz==2025.1
PyYAML==6.0.2
pyzmq==26.2.1
questionary==2.1.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.9.4
rpds-py==0.23.1
rsa==4.9
ruamel.yaml==0.18.5
ruamel.yaml.clib==0.2.7
ruff==0.11.11
safehttpx==0.1.6
safetensors==0.5.3
scikit-image==0.22.0
scikit-learn==1.3.2
scipy==1.11.1
semver==3.0.4
Send2Trash==1.8.3
sentencepiece==0.2.0
sentry-sdk==2.22.0
seqeval==1.2.2
setproctitle==1.3.5
setuptools==80.9.0
shellingham==1.5.4
simple-websocket==1.1.0
simplejson==3.20.1
simsimd==6.2.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.2
soundfile==0.13.1
soupsieve==2.6
spaces==0.37.0
spandrel==0.4.1
stack-data==0.6.3
starlette==0.46.0
strawberry-graphql==0.243.0
stringzilla==3.12.2
swin_window_process==0.0.0
sympy==1.14.0
tabulate==0.9.0
tb-nightly==2.20.0a20250307
tensorboard==2.15.2
tensorboard-data-server==0.7.2
termcolor==2.5.0
terminado==0.18.1
terminaltables==3.1.10
text-unidecode==1.3
threadpoolctl==3.5.0
tifffile==2025.2.18
timm==0.9.7
tinycss2==1.4.0
tokenizers==0.21.1
toml==0.10.2
tomli==2.2.1
tomlkit==0.12.0
torch==2.1.2+cu121
torchaudio==2.1.2+cu121
torchmetrics==1.7.2
torchsde==0.2.6
torchvision==0.16.2+cu121
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
trampoline==0.1.2
transformer_engine==2.4.0
transformer_engine_cu12==2.4.0
transformers==4.50.3
transparent-background==1.2.12
triton==2.1.0
truststore==0.8.0
typer==0.16.0
types-python-dateutil==2.9.0.20241206
typing_extensions==4.14.0
typing-inspect==0.9.0
tzdata==2025.1
uri-template==1.3.0
urllib3==2.4.0
uv==0.7.8
uvicorn==0.34.0
volcengine-python-sdk==1.0.124
voluptuous==0.14.2
wandb==0.20.1
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==11.0.3
Werkzeug==3.1.3
wget==3.2
wheel==0.42.0
wrapt==1.14.1
wsproto==1.2.0
xformers==0.0.23.post1
xxhash==3.5.0
yacs==0.1.8
yapf==0.43.0
yarl==1.18.3
zipp==3.22.0
zope.event==5.0
zope.interface==7.2
zstandard==0.22.0
lpips==0.1.4
torch_fidelity==0.3.0
open-clip-torch==2.32.0
